Modularize: split into sim/render/app, add Player/World abstractions
Reorganize sources under src/{sim,render,app}/ so the simulation tier no
longer pulls in any OpenGL headers. The intent is that a future port to
Godot/Unity/etc. can rewrite src/render and src/app while reusing
src/sim verbatim.

Architecture:
  sim/  - pure C++, no GL. Owns orbital mechanics, satellites, players,
          and the World. Controllers (Human, AI) drive a single Player
          per tick.
  render/ - all OpenGL. SatelliteView holds GL buffers; WorldView pools
            views to match World's satellite count each frame.
  app/  - GLFW glue. Translates input into HumanController state and
          drives the per-frame loop.

Notable changes:
- Replace OrbitalSet with sim::World containing sim::Player(s). Each
  player owns satellites (unique_ptr for stable addresses), tracks its
  own selection, and contributes a dilation vote.
- Time dilation is now sim seconds per real second (double). Effective
  rate is the min over human players' requested rates; default 500.0,
  floor 1.0 (real time).
- Add sim::SatelliteSpec for launch-time attributes (name, mass, fuel,
  max thrust, dV per press, initial r/v).
- Add sim::Controller interface + HumanController and stub AiController
  so input and AI policy share a single seam.
- HUD now reads const sim::World&, draws targeting lines from the
  active player's selected satellite, and shows time/planning status.
- const-correct EarthOrbit and Cube/Ellipse3d construction; no more
  const_cast at the sim/render boundary.
- Add destructors (and delete copy ctors) on Cube/Ellipse3d to stop
  leaking GL handles when satellites are removed.
- Fix uninitialized did_maneuver_, drop unused dump_state() decl,
  remove member-init-order warning, fix shader.h catch-by-value, and
  fix narrowing conversion in writer.cpp.

+++ src/render/writer.cpp
@@ -15,8 +15,8 @@
 #include <ft2build.h>
 #include FT_FREETYPE_H
 // GL includes
-#include "shader.h"
-#include "writer.h"
+#include "render/shader.h"
+#include "render/writer.h"
 
 FtWriter::FtWriter(GLuint WIDTH, GLuint HEIGHT): shader_("shaders/text.vs", "shaders/text.frag") {
 
@@ -77,7 +77,7 @@ FtWriter::FtWriter(GLuint WIDTH, GLuint HEIGHT): shader_("shaders/text.vs", "sha
         texture,
         glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
         glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-        face->glyph->advance.x
+        static_cast<GLuint>(face->glyph->advance.x)
     };
     Characters.insert(std::pair<GLchar, Character>(c, character));
     }
